Issue 4161 - Data Storage Temp Credentials for Azure doesn't work
